requirements.txt
Django==3.0.3
PyYAML==5.3
